Refactor: Move client components next to page
