+++ apps/web/src/app/[locale]/HomeClient.tsx
