+++ apps/web/src/app/[locale]/dish/[id]/MealDetailsClient.tsx
