+++ apps/web/src/app/[locale]/page.tsx
@@ -1,5 +1,5 @@
 import { loadMessages } from "@/lib/internationalisation/i18n";
-import HomeClient from "@/components/Home/HomeClient";
+import HomeClient from "@/app/[locale]/HomeClient";
 import { Locale } from "@calculories/shared-types";
 
 export default async function Page({
