+++ apps/web/src/app/[locale]/settings/page.tsx
@@ -1,6 +1,6 @@
 import { loadMessages } from "@/lib/internationalisation/i18n";
 import { Locale } from "@calculories/shared-types";
-import SettingsClient from "@/components/Settings/SettingsClient";
+import SettingsClient from "@/app/[locale]/settings/SettingsClient";
 
 export default async function Settings({
   params,
